initializing the signal handler in the universe

+++ hotspot/src/share/vm/swap/SwapWriter.h
@@ -15,7 +15,7 @@ class SwapWriter {
 public:
 	SwapWriter();
 	virtual ~SwapWriter();
-	static SSDRange swapOut (void * va, int np, long off);
+	static SSDRange swapOut (void * va, int np, size_t off);
 };
 
 #endif /* SWAPWRITER_H_ */
